完成權限後端

+++ src/main/java/com/delta/auth/Service/impl/BaseServiceImpl.java
@@ -1,12 +1,16 @@
 package com.delta.auth.Service.impl;
 
+import com.baomidou.mybatisplus.annotations.TableId;
 import com.baomidou.mybatisplus.mapper.BaseMapper;
 import com.baomidou.mybatisplus.mapper.EntityWrapper;
-import com.delta.auth.common.ServerResponse;
+import com.delta.common.utils.ServerResponse;
 import lombok.extern.slf4j.Slf4j;
 
+import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
 import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 /**
  * @Classname BaseServiceImpl
@@ -15,13 +19,13 @@ import java.util.List;
  * @Since 1.8
  */
 @Slf4j
-public abstract class BaseServiceImpl<T extends BaseMapper,K> {
+public abstract class BaseServiceImpl<T extends BaseMapper, K> {
 
     protected T dao;
 
     //public BaseServiceImpl(){ }
 
-    public BaseServiceImpl(T dao){
+    public BaseServiceImpl(T dao) {
         this.dao = dao;
     }
 
@@ -35,6 +39,9 @@ public abstract class BaseServiceImpl<T extends BaseMapper,K> {
 
     public ServerResponse<K> deleteOne(K k) {
         EntityWrapper<K> wrapper = getDeleteAndUpdateWrapper(k);
+        if (null == wrapper || wrapper.isEmptyOfWhere()) {
+            return ServerResponse.createByErrorMessage("刪除條件不能為空");
+        }
         int count = dao.delete(wrapper);
         if (count != 1) {
             return ServerResponse.createByErrorMessage("刪除失敗");
@@ -53,22 +60,43 @@ public abstract class BaseServiceImpl<T extends BaseMapper,K> {
     }
 
     public K selectOne(K k) {
-        return (K)dao.selectOne(k);
+        return (K) dao.selectOne(k);
     }
 
     public List<K> selectList() {
         return dao.selectList(new EntityWrapper<>());
     }
 
-    protected abstract EntityWrapper<K> getDeleteAndUpdateWrapper(K k);
+    protected EntityWrapper<K> getDeleteAndUpdateWrapper(K k) {
+        EntityWrapper<K> wrapper = new EntityWrapper<>();
+        Class clazz = k.getClass();
+        Field idField = getField(clazz, TableId.class);
+        Field dataVersion = null;
+        try {
+            dataVersion = clazz.getDeclaredField("dataVersion");
+            dataVersion.setAccessible(true);
+            wrapper.eq("data_version", dataVersion.get(k));
+            if (null == idField) {
+                log.info(String.format("entity %s is not declare id field", clazz.toString()));
+            } else {
+                idField.setAccessible(true);
+                wrapper.eq(humpToLine(idField.getName()), idField.get(k));
+            }
+        } catch (NoSuchFieldException e) {
+            e.printStackTrace();
+        } catch (IllegalAccessException e) {
+            e.printStackTrace();
+        }
+        return wrapper;
+    }
 
-    protected K updateDataVersion(K k){
+    protected K updateDataVersion(K k) {
         Class clazz = k.getClass();
         try {
             Field dataVersion = clazz.getDeclaredField("dataVersion");
             dataVersion.setAccessible(true);
             Integer originDataVersion = Integer.parseInt(dataVersion.get(k).toString());
-            dataVersion.set(k,originDataVersion + 1);
+            dataVersion.set(k, originDataVersion + 1);
         } catch (NoSuchFieldException e) {
             log.info(String.format("entity %s no declaredField of dataVersion", clazz.toString()));
             return k;
@@ -77,5 +105,36 @@ public abstract class BaseServiceImpl<T extends BaseMapper,K> {
             return k;
         }
         return k;
+    }
+
+    private Pattern humpPattern = Pattern.compile("[A-Z]");
+
+    /**
+     * 駝峰轉下劃線
+     *
+     * @param str
+     * @return
+     */
+    private String humpToLine(String str) {
+        Matcher matcher = humpPattern.matcher(str);
+        StringBuffer sb = new StringBuffer();
+        while (matcher.find()) {
+            matcher.appendReplacement(sb, "_" + matcher.group(0).toLowerCase());
+        }
+        matcher.appendTail(sb);
+        return sb.toString();
+    }
+
+    private Field getField(Class clazz, Class annotation) {
+        Field[] fields = clazz.getDeclaredFields();
+        Object id = null;
+        for (Field field : fields) {
+            field.setAccessible(true);
+            id = field.getAnnotation(annotation);
+            if (null != id) {
+                return field;
+            }
+        }
+        return null;
     }
 }
